Deprecate CommandCreateError::Exists
Commands persist between reloads of the plugin. If the (aircraft)
plugin creates a command, it will work on the first run. Reload the
aircraft and it will give a false positive error by finding the
command created on the previous run.

XPLMCreateCommand returns the existing command ref if there is one.

Fixes #15